update Checkbox Submission Component
